move dump and profile metadata to a dedicated package
