+++ pkg/security/security_profile/dump/activity_dump_profile_proto_enc_v1.go
@@ -14,6 +14,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/security/secl/model"
 	"github.com/DataDog/datadog-agent/pkg/security/security_profile"
 	"github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree"
+	mtdt "github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree/metadata"
 )
 
 // ActivityDumpToSecurityProfileProto serializes an Activity Dump to a Security Profile protobuf representation
@@ -25,7 +26,7 @@ func ActivityDumpToSecurityProfileProto(input *ActivityDump) *proto.SecurityProf
 	output := proto.SecurityProfile{
 		Status:   uint32(model.AnomalyDetection + model.AutoSuppression),
 		Version:  security_profile.LocalProfileVersion,
-		Metadata: adMetadataToProto(&input.Metadata),
+		Metadata: mtdt.MetadataToProto(&input.Metadata),
 		Syscalls: input.ActivityTree.ComputeSyscallsList(),
 		Tags:     make([]string, len(input.Tags)),
 		Tree:     activity_tree.ActivityTreeToProto(input.ActivityTree),
