+++ pkg/security/security_profile/dump/activity_dump_proto_dec_v1.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/pkg/security/config"
 	"github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree"
+	mtdt "github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree/metadata"
 )
 
 func protoToActivityDump(dest *ActivityDump, ad *adproto.SecDump) {
@@ -23,7 +24,7 @@ func protoToActivityDump(dest *ActivityDump, ad *adproto.SecDump) {
 	dest.Host = ad.Host
 	dest.Service = ad.Service
 	dest.Source = ad.Source
-	dest.Metadata = ProtoMetadataToMetadata(ad.Metadata)
+	dest.Metadata = mtdt.ProtoMetadataToMetadata(ad.Metadata)
 
 	dest.Tags = make([]string, len(ad.Tags))
 	copy(dest.Tags, ad.Tags)
@@ -31,29 +32,4 @@ func protoToActivityDump(dest *ActivityDump, ad *adproto.SecDump) {
 
 	dest.ActivityTree = activity_tree.NewActivityTree(dest, "activity_dump")
 	activity_tree.ProtoDecodeActivityTree(dest.ActivityTree, ad.Tree)
-}
-
-// ProtoMetadataToMetadata decodes a Metadata structure
-func ProtoMetadataToMetadata(meta *adproto.Metadata) Metadata {
-	if meta == nil {
-		return Metadata{}
-	}
-
-	return Metadata{
-		AgentVersion:      meta.AgentVersion,
-		AgentCommit:       meta.AgentCommit,
-		KernelVersion:     meta.KernelVersion,
-		LinuxDistribution: meta.LinuxDistribution,
-		Arch:              meta.Arch,
-
-		Name:              meta.Name,
-		ProtobufVersion:   meta.ProtobufVersion,
-		DifferentiateArgs: meta.DifferentiateArgs,
-		Comm:              meta.Comm,
-		ContainerID:       meta.ContainerId,
-		Start:             activity_tree.ProtoDecodeTimestamp(meta.Start),
-		End:               activity_tree.ProtoDecodeTimestamp(meta.End),
-		Size:              meta.Size,
-		Serialization:     meta.GetSerialization(),
-	}
 }
