+++ pkg/security/security_profile/dump/activity_dump_proto_enc_v1.go
@@ -12,6 +12,7 @@ import (
 	adproto "github.com/DataDog/agent-payload/v5/cws/dumpsv1"
 
 	"github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree"
+	mtdt "github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree/metadata"
 )
 
 func activityDumpToProto(ad *ActivityDump) *adproto.SecDump {
@@ -25,7 +26,7 @@ func activityDumpToProto(ad *ActivityDump) *adproto.SecDump {
 		Service: ad.Service,
 		Source:  ad.Source,
 
-		Metadata: adMetadataToProto(&ad.Metadata),
+		Metadata: mtdt.MetadataToProto(&ad.Metadata),
 
 		Tags: make([]string, len(ad.Tags)),
 		Tree: activity_tree.ActivityTreeToProto(ad.ActivityTree),
@@ -34,30 +35,4 @@ func activityDumpToProto(ad *ActivityDump) *adproto.SecDump {
 	copy(pad.Tags, ad.Tags)
 
 	return pad
-}
-
-func adMetadataToProto(meta *Metadata) *adproto.Metadata {
-	if meta == nil {
-		return nil
-	}
-
-	pmeta := &adproto.Metadata{
-		AgentVersion:      meta.AgentVersion,
-		AgentCommit:       meta.AgentCommit,
-		KernelVersion:     meta.KernelVersion,
-		LinuxDistribution: meta.LinuxDistribution,
-
-		Name:              meta.Name,
-		ProtobufVersion:   meta.ProtobufVersion,
-		DifferentiateArgs: meta.DifferentiateArgs,
-		Comm:              meta.Comm,
-		ContainerId:       meta.ContainerID,
-		Start:             activity_tree.TimestampToProto(&meta.Start),
-		End:               activity_tree.TimestampToProto(&meta.End),
-		Size:              meta.Size,
-		Arch:              meta.Arch,
-		Serialization:     meta.Serialization,
-	}
-
-	return pmeta
 }
