+++ pkg/security/security_profile/profile/profile_proto_dec_v1.go
@@ -14,7 +14,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/security/secl/model"
 	"github.com/DataDog/datadog-agent/pkg/security/security_profile"
 	"github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree"
-	"github.com/DataDog/datadog-agent/pkg/security/security_profile/dump"
+	mtdt "github.com/DataDog/datadog-agent/pkg/security/security_profile/activity_tree/metadata"
 )
 
 func protoToSecurityProfile(output *SecurityProfile, input *proto.SecurityProfile) {
@@ -27,7 +27,7 @@ func protoToSecurityProfile(output *SecurityProfile, input *proto.SecurityProfil
 	if input.Version == security_profile.LocalProfileVersion {
 		output.autolearnEnabled = true
 	}
-	output.Metadata = dump.ProtoMetadataToMetadata(input.Metadata)
+	output.Metadata = mtdt.ProtoMetadataToMetadata(input.Metadata)
 
 	output.Tags = make([]string, len(input.Tags))
 	copy(output.Tags, input.Tags)
